clean, fix: _baseSqlQueryHandler removed, sqlQueryRun used explicitly
Numerous unspecific problems with `sqlite` database appeared when trying to shift the responsibility of handling async context to _baseQueryHandler, so I got rid of it and commit to work more with async context directly (although it looks hideous imho).

The problem where text in annotations or notes contained single quotes and would be interpreted as string closing quotes in a sql query. These are now automatically escaped.